refactor: change useFeatureFlag import path
